+++ src/client/src/AppTheme.ts
@@ -3,7 +3,7 @@ import { createTheme } from "@mui/material/styles";
 export const theme = createTheme({
   palette: {
     primary: {
-      main: "#212121",
+      main: "#1c2834",
       contrastText: "#ffffff",
     },
     secondary: {
@@ -28,6 +28,9 @@ export const theme = createTheme({
       main: "#0b7285",
       secondary: "#a65462",
     },
+    boxShadow: "rgba(0, 0, 0, 0.17)",
+    background: { default: "#ffffff", lightgrey: "#f1f3f5", darkgrey: "#787878", dark: "rgba(0, 0, 0, 0.5)" },
+    buttonSelected: "#AA3526",
   },
   typography: {
     fontFamily: "Lato",