+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 
 - Added borehole geometry panel.
 
+### Changed
+
+- Added `outcrop` to codelist `borehole_type`.
+- Made boreholes table on the map page collapsible.
+- Changed sidebar layout and design.
+
 ### Fixed
 
 - Fixed precisions where not correctly displayed for imported coordinates.